🧑‍💻 Combine G60-G61
