+++ Marlin/src/module/motion.h
@@ -48,12 +48,6 @@ extern bool relative_mode;
 extern xyze_pos_t current_position,  // High-level current tool position
                   destination;       // Destination for a move
 
-// G60/G61 Position Save and Return
-#if SAVED_POSITIONS
-  extern Flags<SAVED_POSITIONS> did_save_position;
-  extern xyze_pos_t stored_position[SAVED_POSITIONS];
-#endif
-
 // Scratch space for a cartesian result
 extern xyz_pos_t cartes;
