+++ buildroot/share/PlatformIO/scripts/preflight-checks.py
@@ -149,12 +149,16 @@ if pioutil.is_pio_build():
         # Check for old files indicating an entangled Marlin (mixing old and new code)
         #
         mixedin = []
-        p = project_dir / "Marlin/src/lcd/dogm"
+        p = mpath / "src/lcd/dogm"
         for f in [ "ultralcd_DOGM.cpp", "ultralcd_DOGM.h", "u8g_dev_ssd1306_sh1106_128x64_I2C.cpp", "u8g_dev_ssd1309_12864.cpp", "u8g_dev_st7565_64128n_HAL.cpp", "u8g_dev_st7920_128x64_HAL.cpp", "u8g_dev_tft_upscale_from_128x64.cpp", "u8g_dev_uc1701_mini12864_HAL.cpp", "ultralcd_st7920_u8glib_rrd_AVR.cpp" ]:
             if (p / f).is_file():
                 mixedin += [ f ]
-        p = project_dir / "Marlin/src/feature/bedlevel/abl"
+        p = mpath / "src/feature/bedlevel/abl"
         for f in [ "abl.cpp", "abl.h" ]:
+            if (p / f).is_file():
+                mixedin += [ f ]
+        f = mpath / "src/gcode/feature/pause"
+        for f in [ "G60.cpp", "G61.cpp" ]:
             if (p / f).is_file():
                 mixedin += [ f ]
         if mixedin:
