+++ ini/features.ini
@@ -318,7 +318,7 @@ HAS_ZV_SHAPING                         = build_src_filter=+<src/gcode/feature/in
 GCODE_MACROS                           = build_src_filter=+<src/gcode/feature/macro>
 GRADIENT_MIX                           = build_src_filter=+<src/gcode/feature/mixing/M166.cpp>
 NONLINEAR_EXTRUSION                    = build_src_filter=+<src/gcode/feature/nonlinear>
-HAS_SAVED_POSITIONS                    = build_src_filter=+<src/gcode/feature/pause/G60.cpp> +<src/gcode/feature/pause/G61.cpp>
+HAS_SAVED_POSITIONS                    = build_src_filter=+<src/gcode/feature/pause/G60_G61.cpp>
 PARK_HEAD_ON_PAUSE                     = build_src_filter=+<src/gcode/feature/pause/M125.cpp>
 FILAMENT_LOAD_UNLOAD_GCODES            = build_src_filter=+<src/gcode/feature/pause/M701_M702.cpp>
 HAS_STEALTHCHOP                        = build_src_filter=+<src/gcode/feature/trinamic/M569.cpp>
